Change metricsQueryType in PlannedStmt from int to int8
Int8 for metricsQueryType is big enough, int is not necessary

+++ src/include/nodes/plannodes.h
@@ -148,7 +148,7 @@ typedef struct PlannedStmt
 	/* 
  	 * GPDB: whether a query is a SPI inner query for extension usage 
  	 */
-	uint		metricsQueryType;
+	int8		metricsQueryType;
 } PlannedStmt;
 
 /*
